requirements.txt
fastapi
uvicorn
motor
python-dotenv
dependency-injector
email-validator